Increase the SEC/PEI ram size for Q35 (#82)
## Description

With the transition to 64bit PEI, a few hob creators naturally start to
produce larger HOBs. This could cause certain build configuration to
fail to create hob before the memory becomes available.

This change doubled the size for temporary RAM to accommodate the memory
usage.

- [x] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

After the change, the allocation failure that occurred when the
performance build flag was enabled no longer happens.

## Integration Instructions

N/A